refactor: move mocks files to mocks folder to be shared with whole application
